+++ NEWS.md
@@ -1,3 +1,7 @@
+# PlackettLuce (development version)
+
+* Avoid computing variance-covariance matrix in `predict.PLADMM(vcov = FALSE)` and `AIC` when new data specified (partial fix to #50).
+
 # PlackettLuce 0.4.0
 
 * New `pladmm` function to fit the Plackett-Luce model with log-worth modelled by item covariates.